requirements.txt
openai>=1.0.0
ast-tools>=0.1.0
bandit>=1.7.0
flake8>=6.0.0
pylint>=2.17.0
click>=8.1.0
pyyaml>=6.0
colorama>=0.4.6
rich>=13.0.0
gitpython>=3.1.0
pathspec>=0.11.0
